msm: camera: fix AXI RESET cmd field to be 32 byte
In 32 bit userspace and 64 bit kernel,
using long could cause the bit lenght difference in kernel and userspace.
change to use uint32_t instead of long.

+++ include/media/msmb_isp.h
@@ -219,7 +219,7 @@ struct msm_vfe_axi_halt_cmd {
 
 struct msm_vfe_axi_reset_cmd {
 	uint32_t blocking; //Boolean whether stop camif is to be done
-	unsigned long frame_id;
+	uint32_t frame_id;
 };
 
 struct msm_vfe_axi_restart_cmd {
